energy_neural_system/requirements.txt
numpy
matplotlib
Pillow
opencv-python
mss
tk
networkx
pyvis 